Test.DocTest

Documentation

doctest :: [String] -> IO ()

Run doctest with given list of arguments.

Example:

>>> doctest ["-iexample/src", "example/src/Example.hs"]
...
Examples: 2  Tried: 2  Errors: 0  Failures: 0

This can be used to create a Cabal test suite that runs doctest for your project.

If a directory is given, it is traversed to find all .hs and .lhs files inside of it, ignoring hidden entries.