Test.DocTest.Internal.Parse

Documentation

data Module a

Documentation for a module grouped together with the modules name.

Constructors

Module

Fields

moduleName :: String

moduleSetup :: Maybe a

moduleContent :: [a]

data DocTest

Constructors

Example Expression ExpectedResult

Property Expression

type Expression = String

type ExpectedResult = [ExpectedLine]

data ExpectedLine

Constructors

ExpectedLine [LineChunk]

WildCardLine

data LineChunk

Constructors

LineChunk String

WildCardChunk

extractDocTests

Arguments

:: [String]

-> IO [Module [Located DocTest]]
Extracted DocTests

Extract DocTests from all given modules and all modules included by the given modules.

extractDocTests = fmap parseModules . extract

parseModules :: [Module (Located String)] -> [Module [Located DocTest]]