Test.DocTest.Internal.Run

Documentation

type Result = Summary

data Config

Constructors

Config

Fields

ghcOptions :: [String]

fastMode :: Bool

preserveIt :: Bool

verbose :: Bool

repl :: (String, [String])

isSuccess :: Result -> Bool

doctest :: [String] -> IO ()

Run doctest with given list of arguments.

Example:

>>> doctest ["-iexample/src", "example/src/Example.hs"]
...
Examples: 2  Tried: 2  Errors: 0  Failures: 0

This can be used to create a Cabal test suite that runs doctest for your project.

If a directory is given, it is traversed to find all .hs and .lhs files inside of it, ignoring hidden entries.

defaultConfig :: Config

doctestWithRepl :: (String, [String]) -> [String] -> IO ()

doctestWith :: Config -> IO ()

data Summary

Summary of a test run.

Constructors

Summary

Fields

sExamples :: Int

sTried :: Int

sErrors :: Int

sFailures :: Int

evaluateResult :: Result -> IO ()

doctestWithResult :: Config -> IO Result

runDocTests :: Config -> [Module [Located DocTest]] -> IO Result